rename basic-interactor -> adder-interactor, templating fix
