+++ contracts/examples/adder/interactor/Cargo.toml
@@ -1,16 +1,16 @@
 [package]
-name = "basic-interactor"
+name = "adder-interactor"
 version = "0.0.0"
 authors = ["MultiversX <contact@multiversx.com>"]
 edition = "2024"
 publish = false
 
 [[bin]]
-name = "basic-interactor"
-path = "src/basic_interactor_main.rs"
+name = "adder-interactor"
+path = "src/adder_interactor_main.rs"
 
 [lib]
-path = "src/basic_interactor.rs"
+path = "src/adder_interactor.rs"
 
 [dependencies.adder]
 path = ".."
