+++ contracts/examples/adder/interactor/src/adder_interactor_main.rs
@@ -1,0 +1,6 @@
+extern crate adder_interactor;
+
+#[tokio::main]
+pub async fn main() {
+    adder_interactor::adder_cli().await;
+}
