--- contracts/examples/adder/interactor/src/basic_interactor_main.rs
@@ -1,6 +1,0 @@
-extern crate basic_interactor;
-
-#[tokio::main]
-pub async fn main() {
-    basic_interactor::adder_cli().await;
-}
